+++ server/src/app.module.ts
@@ -7,6 +7,7 @@ import { UsersModule } from './users/users.module';
 import { ImagesModule } from './images/images.module';
 import { CreditsModule } from './credits/credits.module';
 import { IAPModule } from './iap/iap.module';
+import { CategoriesModule } from './categories/categories.module';
 import { ConfigModule } from '@nestjs/config';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
@@ -34,8 +35,9 @@ import { join } from 'path';
     ImagesModule,
     CreditsModule,
     IAPModule,
+    CategoriesModule,
   ],
   controllers: [AppController],
   providers: [AppService],
 })
-export class AppModule {}
+export class AppModule { }